Implemented bulk import

+++ server/create_database.sql
@@ -63,7 +63,7 @@ CREATE TABLE Inventories(
 	id int not null AUTO_INCREMENT,
 	ingredient_id int not null,
 	package_type enum('sack', 'pail', 'drum', 'supersack', 'truckload', 'railcar') not null,
-	num_packages int not null default 0,
+	num_packages double not null default 0,
 
 	FOREIGN KEY (ingredient_id) REFERENCES Ingredients(id),
 	PRIMARY KEY (id)
